fix:[CI-13143] refatoring test infrastructure
