+++ convert/jenkinsjson/json/findFiles_test.go
@@ -1,49 +1,16 @@
 package json
 
 import (
-	"encoding/json"
-	"os"
-	"path/filepath"
 	"testing"
 
 	harness "github.com/drone/spec/dist/go"
 	"github.com/google/go-cmp/cmp"
 )
 
-type runner struct {
-	name  string
-	input Node
-	want  *harness.Step
-}
-
-func prepare(t *testing.T, filename string, step *harness.Step) runner {
-
-	workingDir, err := os.Getwd()
-	if err != nil {
-		t.Fatalf("failed to get current working directory: %v", err)
-	}
-
-	jsonData, err := os.ReadFile(filepath.Join(workingDir, "../convertTestFiles/findFiles", filename+".json"))
-	if err != nil {
-		t.Fatalf("failed to read JSON file: %v", err)
-	}
-
-	var inputNode Node
-	if err := json.Unmarshal(jsonData, &inputNode); err != nil {
-		t.Fatalf("failed to decode JSON: %v", err)
-	}
-
-	return runner{
-		name:  filename,
-		input: inputNode,
-		want:  step,
-	}
-}
-
 func TestConvertFindFiles(t *testing.T) {
 
 	var tests []runner
-	tests = append(tests, prepare(t, "findFiles_GlobOnly", &harness.Step{
+	tests = append(tests, prepare(t, "findFiles/findFiles_GlobOnly", &harness.Step{
 		Id:   "findFiles4d4efe",
 		Name: "findFiles",
 		Type: "plugin",
@@ -54,7 +21,7 @@ func TestConvertFindFiles(t *testing.T) {
 			},
 		},
 	}))
-	tests = append(tests, prepare(t, "findFiles_Excludes", &harness.Step{
+	tests = append(tests, prepare(t, "findFiles/findFiles_Excludes", &harness.Step{
 		Id:   "findFilescc58db",
 		Name: "findFiles",
 		Type: "plugin",
